qnd: Changed the fix of the entry_media_type_id data type fix - since it was originally smallint(6) as it was defined in the fact. We changed the dimension (dwh_dim_entry_media_type) to int(11) because the aggregation table (dwh_hourly_events_devices) was defined with an entry_media_type_id of the int(11) data type (this was not good). We now changed the aggregation to smallint(6) and the dimension as well (since modifying a data type in the fact is too long to do - and unnecessary in this case)
git-svn-id: svn+ssh://kelev.kaltura.com/usr/local/kalprod/dwh/core/trunk@7090 09e43ab6-4768-4770-81d3-4e3a7c8fdb74

+++ ddl/dw/aggr/dwh_hourly_events_devices.sql
@@ -11,7 +11,7 @@ CREATE TABLE kalturadw.`dwh_hourly_events_devices` (
   `os_id` INT NOT NULL DEFAULT -1,
   `browser_id` INT NOT NULL DEFAULT -1,
   `ui_conf_id` INT NOT NULL DEFAULT -1,
-  `entry_media_type_id` INT(11) NOT NULL DEFAULT -1,
+  `entry_media_type_id` SMALLINT(6) NOT NULL DEFAULT -1,
   `sum_time_viewed` DECIMAL(20,3) DEFAULT NULL,
   `count_time_viewed` INT DEFAULT NULL,
   `count_plays` INT DEFAULT NULL,
